Move tests to separated files
